Log more information when nylas has a request error
This ensures we log out the full error object, the status code and
message, and the full response body when there is an error in the
`NylasConnection#request` method. This is to work around a bug where we
are only getting a non-descript error from the nylas api, and we need
more information to suss out.