Fix raw sending to parse message correctly
Summary: Apparently, the `JSON.parse()` I deleted had a purpose! We do need to parse JSON ourselves in one case—if we flag a request as `json: false`. That happens when we send a raw message because we want to send raw MIME (not JSON), but we want to parse the JSON response we get into a `Message` object.

Test Plan: Tested locally.

Reviewers: evan, spang, khamidou, mark

Reviewed By: mark

Subscribers: mark

Differential Revision: https://phab.nylas.com/D5859