Delete references to Tag
Summary: We still reference `Tag` in several places, and that hasn't been a part of our API for a long, long time. I deleted that model and the references to `Tag` objects in `Thread`. Instead, users can just use the normal `save()` function to update an object.

Test Plan: Tested locally.

Reviewers: spang, khamidou, evan

Reviewed By: spang

Differential Revision: https://phab.nylas.com/D5881