Support views for counts and IDs
Summary: A customer reported an issue with the Node SDK, where calling `list` with `view: 'count'` threw an error because `_getModelCollection` always expects an array, not a number. We have a method `count`, but I've added more helpful error messages for the methods that shouldn't support `count`, and I've changed `getModelCollection` to `getItems` because `view: 'ids'` returns an array of IDs.

Test Plan: Tested.

Reviewers: spang, khamidou, evan

Reviewed By: spang

Differential Revision: https://phab.nylas.com/D5901