Fix handling of error when not a JSON request
When sending raw MIME emails, the `json` request option is `false`, which means that the body is still a stringified JSON instead of an object. This means that when ever there was a failure, the error object never included any information about what happened because it was referencing a `.message` property on a string.