Separate event-participant and email-participant
Summary:
Previously, the participant class referred exclusively to event participants and the contact class had 2 fields - email and name, and was used for the messsage fields pretaining to an email participant (to, from, cc, bcc etc). When I expanded contact model to include the new v2 fields, this caused problems.

This diff renames participants to event-participants and renames what used to be contact to email-participants. This makes it more clear where the different participants are being used and frees up `contact` for all the new fields.

Test Plan: manual

Reviewers: halla, khamidou

Reviewed By: halla

Differential Revision: https://phab.nylas.com/D6113