Create and use AttributeDateYYYYMMDD for birthday
Summary:
The existing Date attribute is meant to be used for message time stamps where `toJSON` converts the passed in date to a time string and `fromJSON` converts from a string to a Date in time format in seconds. This isnt very helpful for birthdays where the date in YYYY-MM-DD is more important than any time / timezone info.

This diff creates a new Attribute called `AttributeDateYYYYMMDD` and then uses this attribute for contact birthdays.

Test Plan: manual

Reviewers: halla, khamidou

Reviewed By: khamidou

Differential Revision: https://phab.nylas.com/D6147