Improve mismatched version messaging
Summary:
I wanted to improve the warnings for if the version of the api that the sdk supported differed from the the version of the api that the application was running on. Initially we had one simple warning message to update the version of the sdk. While this is likely the most common condition, this diff distinguishes between the case where the sdk needs to be updated and the case where the application needs to update its api version through the dashboard.

Also, I refactored the warning for mismatched versioning to be a function outside of the request/response. The main benefit of this is that it makes testing easier. It also makes the response function shorter and more clear.

I also added unit tests for the warnings.

Test Plan: unit tests

Reviewers: mark, halla, khamidou

Reviewed By: mark

Differential Revision: https://phab.nylas.com/D6166