Build out threads example
Summary: For the sample app, I added some functions and views for the threads example. Previously, when you linked an account, you would be automatically sent to a screen with the first thread for that account. I didnt think this was very clear for a user who was testing it out. Thus I had the user click a link to get this and wrapped it in a border. Also added a list example.

Test Plan: manual

Reviewers: mike, mark, halla, khamidou

Reviewed By: mike, khamidou

Subscribers: khamidou, tomasz

Differential Revision: https://phab.nylas.com/D6199