[60372] Return better error details from API (#236)
# Description
Previously, we would only return generally the error message. Now we return a custom error object (that extends `Error`) that properly stores all of the error values we get back from the API and sets them in a way that is more native than just returning a string.

# Usage
The new `NylasApiError` class extends the `Error` class and both sets the build-in properties of Error and extends the error class to provide more values if needed. The following properties are made availalbe by the class:

- `name` (string): The name of the error, derived from the [API Errors page](https://developer.nylas.com/docs/developer-tools/api/errors/) (e.g. `Bad Request`)
- `message` (string): The error message returned from the API (e.g. `No recipients specified`)
- `statusCode` (number): The status code returned from the API call (e.g. `400`)
- `type` (string): The type of error returned from the API (e.g. `invalid_request_error`)
- `missingFields` (string[]): The array of fields that were missing in the call, if applicable (e.g. `["name", "email"]`)
- `serverError` (string): The error returned by the provider (e.g. `interaction_required`)
- `stack` (string): The stacktrace from JavaScript