[62051] Draft tracking object not being added to a pre-existing `draft` object (#241)
The Node SDK only added a tracking object to the payload if the draft was never "created" before (created in the Nylas API, as in the object was not `.save()` yet and does not have a draft ID), which is allowed according to the [Message Tracking documentation](https://developer.nylas.com/docs/developer-tools/webhooks/message-tracking/#enabling-tracking). However if the user has already created a `draft` object in the Nylas and tries to send it with a tracking object, it never gets added. This PR allows for a tracking object to be included in either case.