Initial commit of properties refactor
A followup to the previous commit. Now we can pass in a properties object to properly instantiate an object. The typing is also now fixed as now required properties are no longer optional. We have set them to a default value to allow the users to choose how to create an object - either via passing in a properties object or by setting each field manually.