[66695] Refactor Restfulmodel (#267)
As the SDK coverage grows, we need to decouple the standard functionality of a model from the specific functionality/structure of a "classic" Nylas API model. A "classic" Nylas API model is a model used for endpoints like Event, Calendar, Message, etc. where we expect the object to have an id, account_id, and object field. There are a lot of "dumb" models like Contact.Emails that would benefit from being cast as a basic model versus the current RestfulModel, as well as having the ability in the future for creating different types of models.