[64477] Fix readonly Calendar attributes (#276)
This PR addresses an issue that causes some PUT calls for Calendar objects to break. The issue was found when a customer was unable to update an event on API 2.2 with the API rejecting the call because message_id was present in the body. The messageId property is supposed to be readonly. We have 5 properties in Calendar that are now marked as readonly: messageId, owner, iCalUID, status, and jobStatusId