[75824] Job Status improvements (#295)
This PR improves the Job Status support on this SDK. Some models are missing a job_status_id key, as well as the Job Status model itself was missing an original_data key. A new helper method was also introduced to the JobStatus object called isSuccessful() that returns a boolean value representing a successful job status.