[77455] Prevent `Event` from sending empty notification array when notifications were not set (#298)
Some recurring events throw an error from the API when sending an empty list of notifications. This could occur without the user setting an empty list of notifications as the Node SDK defaults all array-type attributes to an empty array even if not directly set the attribute. This PR overrides this functionality and prevents notifications from being set if the user did not explicitly set the attribute.