Updating exchange token response to match API (#279)
A customer noticed that we were returning only the access token in the exchange token response. This doesn't match the API's response so I updated the response body to include the full response body from the call we make to exchange the token. This will return identifying information about the account as well.