[66699] Node SDK v6.0.0 - major refactor release (#302)
This PR represents the major update v6.0.0 of the Node SDK. The main theme of this major release is refactoring and improving the SDK to make it more Typescript-like as well as improving the usability of the SDK by providing better and stricter type hinting as well as expected parameters. This major release can be broken down into three main sections: refactoring RestfulModel, improving deserialization and object creation, and improving the typing of the SDK in general.