[78923] Fix `saveRequestBody()` for `Model` class (#309)
saveRequestBody() was introduced prior to the v6.0 release to enforce read-only attributes not being serialized for outgoing API calls. With the introduction of the Model superclass in v6.0 a lot of objects got refactored from extending RestfulModel to Model, but saveRequestBody() remained on RestfulModel. This broke the serialization of Model properties. To resolve this, we moved saveRequestBody() to the Model superclass.