Add new property to allow for directly adding file_ids
This will allow for users who want to attach files to a draft via fileIds. Due to the existence of `fileIds()` function in the `Message` superclass, we can't set this variable to a friendlier name without causing a breaking change.

This variable only adds files to a draft, to remove a file it must be done on the files array. Upon saving we will make sure to merge unique values in the outgoing payload.