[79908] Add new field in `Draft` for adding files by file IDs (#319)
This PR adds support for adding a new field, fileIdsToAttach: string[], in Draft which will allow users who want to attach files to a draft via file IDs. Due to the existence of fileIds() function in the Message superclass, we can't set this variable to a friendlier name without causing a breaking change.

Please note that this variable only adds files to a draft, to remove a file it must be done on the files array. Upon saving we will make sure to merge unique values in the outgoing payload.