Add support for getting raw messages directly from the API (#332)
This PR enables support for directly getting a raw message from the API. Currently on the SDK you would need to get a message first then run getRaw() on it, which results in 2 API calls. This function exists on MessageRestfulModelCollection allowing you to get the message as raw directly using a message ID resulting in just one API call.