Fix Outbox JSON Parsing Error  (#339)
The outbox send method met a specific condition in nylas-connection where it would return undefined as opposed to the parsed json of the response. This would have the downstream base model throw an error when it tried to work with the json. It was meeting that condition because of an equality check with response.size and 0. However with node-fetch response.size cannot be used/relied on.