Remove use of `.clone` in `node-fetch` operations (#358)
clone on node-fetch < 3.0.0 does not allow you to set highWaterMark which causes issues with fetching large amounts of data after cloning. Since we can't upgrade to node-fetch 3.x at the moment, we shouldn't use it, and instead we should just fetch the buffer as a string and try to parse the JSON from it.