Abstracting middleware logic (#382)
This PR's main change is pulling out the underlying logic of the routes used in the middleware created for Express and moving it to Routes. This new module's purpose is to allow implementing clients that don't use one of the supported frameworks served by ServerBindings still an easy way to integrate Nylas into their code and pair it with our frontend SDKs we provide (@nylas/nylas-js, @nylas/nylas-react).

This code also brings forward a breaking change, the Nylas.exchangeCodeForToken now expects the optional callback parameter to anticipate a AccessToken object instead of a string for the accessToken parameter.