Backport v6.5.0 and v6.5.1 features to v7 branch (#395)
This PR cherry picks the diff commits from main and the v7 branch to ensure beta 1 has all the up to date features and bugfixes:

* Fixed json value for `confirmationEmailToHost` (#377)

This PR fixes the json value mapped to confermationEmailToHost as "email" is actually plural. Adjusting the actual model variable name will be introduced in v7 as it's a breaking change.

* Add `interval_minutes` field in Scheduler booking config (#378)

This PR adds the intervalMinutes field in ScheduleBooking. Note that this field only takes a non-zero, positive integer that's divisible by 5. Else we have validation that runs when attempting to save that will throw an error.

* Fix typing issues in event and scheduler open hours (#381)

Fix issues relating to typing in:
- `Event`: `originalStartTime` should deserialize to `Date`, so replace `number` type to `Date`.
- `SchedulerBookingOpenHours`: The `account_id` filed from the API was being deseralized to `dropdownOptions` instead of `accountId`.

* Add `metadata` field to `JobStatus` (#379)

This PR adds a new metadata field to the JobStatus class.

* v6.5.0 Release (#383)

New nylas v6.5.0 release provides the following changes

* Fix issue with `EventParticipant` not sending status on new event creation (#386)

This PR ensures that on a create the event's participant status is sent.

* Add additional Event fields (#387)

Add the following fields to `Event`:
* customeEventId: string
* oragnizerEmail: string (read only)
* organizerName: string (read only)
* visibility: string (read only)

* Add codecov support (#388)

* Add codecov support

* Add codecov shield to readme

* Pretty-fy the readme

* v6.5.1 Release (#391)

New nylas v6.5.1 release