Officially support Node >= 16 (#428)
This PR adds an engine entry for Node >= 16 as moving forward we are only supporting frameworks and languages that are currently supported. This PR also removes some code that was backwards compatible with Node < 10 and updates the github actions to only run on the supported LTS versions.

This PR also sets up actions for changes incoming to the v7-staging branch, and removes webhook tunnel support as it will not be available on launch.