Get SDK version reliably (#480)
Because of how we are building our SDK now with the ES Module addition in #477 it has caused importing the `package.json` file to be a bit more tricky given that it no longer exists one directory up. Instead we just generate a version file post-build and during the npm version call and we pull the version from that file. The file is written, and sources the version from, the package.json file.